SpringSecurity Added Authentication for ReviewEndPoints

+++ src/main/java/com/cineLog/cineLog/service/ReviewEntryService.java
@@ -1,8 +1,6 @@
 package com.cineLog.cineLog.service;
 
 
-
-
 import com.cineLog.cineLog.entity.ReviewEntity;
 import com.cineLog.cineLog.entity.UserEntity;
 import com.cineLog.cineLog.repository.ReviewEntityRepo;
@@ -31,35 +29,34 @@ public class ReviewEntryService {
     private UserEntryService userEntryService;
 
 
-
     public List<ReviewEntity> getAll() {
         return reviewEntityRepo.findAll();
     }
 
 
-    public Optional<ReviewEntity> findById(String id){
+    public Optional<ReviewEntity> findById(String id) {
         return reviewEntityRepo.findById(id);
     }
 
 
-
     @Transactional
-    public void saveEntry(ReviewEntity reviewEntity, String username) {
+    public void saveNewUser(ReviewEntity reviewEntity, String username) {
         // Fetch the user by username
         try {
             UserEntity user = userEntryService.findByusername(username);
             reviewEntity.setCreatedAt(LocalDateTime.now());
             ReviewEntity saved = reviewEntityRepo.save(reviewEntity);
             user.getReviewEntities().add(saved);
-            userEntryService.saveEntry(user);
-        }catch (Exception e){
-         //   System.out.println(e);
-            throw new RuntimeException("An error occurred while saving",e);
+            //userEntryService.saveEntry(user);
+            userEntryService.saveNewUser(user);
+        } catch (Exception e) {
+            //   System.out.println(e);
+            throw new RuntimeException("An error occurred while saving", e);
         }
 
     }
 
-    public void saveEntry(ReviewEntity reviewEntity) {
+    public void saveNewUser(ReviewEntity reviewEntity) {
         reviewEntityRepo.save(reviewEntity);
     }
 
@@ -68,11 +65,29 @@ public class ReviewEntryService {
     }
 
 
+    @Transactional
+    public boolean deleteById(String id, String username) {
+        boolean removed = false;
 
-    public void deleteById(String id, String username) {
-        UserEntity user = userEntryService.findByusername(username);
-        user.getReviewEntities().removeIf(x->x.getId().equals(id));
-        userEntryService.saveEntry(user);
-        reviewEntityRepo.deleteById(id);
+        try {
+            UserEntity user = userEntryService.findByusername(username);
+            removed = user.getReviewEntities().removeIf(x -> x.getId().equals(id));
+            if (removed) {
+                userEntryService.saveEntry(user);
+                reviewEntityRepo.deleteById(id);
+            }
+
+
+        } catch (Exception e) {
+            System.out.println(e);
+            throw new RuntimeException("An error occurred while deleting the entry", e);
+        }
+        return removed;
     }
+
+    public UserEntity findByUserName(String userName) {
+        return userEntityRepo.findByusername(userName);
+
+    }
+
 }
